Fix a problem with bad regex and generalize fix for folder resources.
Don't remove the port number from the URL, but replace the colon with an underscore. If we get a request for a cache filename for a URL that ends in a slash, it's a folder, so make a special file just for that folder, called "_index_". Otherwise the file inhibits caching any files inside that folder.